+++ packages/create-vue-vine/template/code/base/src/main.ts
@@ -1,8 +1,0 @@
-import { createApp } from 'vue'
-import { App as VineApp } from './app.vine'
-
-import './styles'
-
-const app = createApp(VineApp)
-
-app.mount('#app')