+++ packages/create-vue-vine/template/code/css/base/src/styles/index.ts
@@ -1,2 +1,0 @@
-import './main.css'
-import './atom.css'