+++ packages/create-vue-vine/template/vite/unocss/src/styles/index.ts
@@ -1,0 +1,2 @@
+import 'virtual:uno.css'
+import './main.css'