+++ packages/create-vue-vine/template/vite/unocss/vite.config.ts
@@ -1,0 +1,17 @@
+import { fileURLToPath } from 'node:url'
+import UnoCSS from 'unocss/vite'
+import { defineConfig } from 'vite'
+import { VineVitePlugin } from 'vue-vine/vite'
+
+export default defineConfig({
+  plugins: [
+    VineVitePlugin(),
+    UnoCSS(),
+  ],
+  resolve: {
+    alias: {
+      '@': fileURLToPath(new URL('./src', import.meta.url)),
+    },
+  },
+})
+