+++ packages/rspack-loader/CHANGELOG.md
@@ -1,9 +1,27 @@
 # @vue-vine/rspack-loader
 
-## 1.7.11-beta.1
+## 1.7.13-beta.1
 
 ### Patch Changes
 
-- Refactor and ready to release (beta) Rspack support.
+- Support initializing project with Rsbuild in scaffold CLI.
+- Updated dependencies
+  - @vue-vine/compiler@1.7.13
+
+## 1.7.12-beta.1
+
+### Patch Changes
+
+- Support rsbuild plugin and add export entry in main package.
+- Updated dependencies
+  - @vue-vine/compiler@1.7.12
+
+## 1.7.11-beta.2
+
+- Support Rsbuild plugin.
+
+## 1.7.11-beta.1
+
+- Refactor common logic and ready to release (beta) Rspack support.
 - Updated dependencies
   - @vue-vine/compiler@1.7.11